Consolidating the channels settings
