+++ src/aslm/view/main_window_content/camera_tab.py
