+++ src/aslm/view/main_window_content/stage_tab.py
