+++ src/aslm/view/main_window_content/tiling_wizard_popup.py
